+++ sdk/storage/storage-file-datalake/src/DataLakeServiceClient.ts
@@ -12,9 +12,13 @@ import { StorageSharedKeyCredential } from "./credentials/StorageSharedKeyCreden
 import { DataLakeFileSystemClient } from "./DataLakeFileSystemClient";
 import {
   FileSystemItem,
+  FileSystemRenameResponse,
   ServiceGenerateAccountSasUrlOptions,
   ServiceListFileSystemsOptions,
-  ServiceListFileSystemsSegmentResponse
+  ServiceListFileSystemsSegmentResponse,
+  ServiceRenameFileSystemOptions,
+  ServiceUndeleteFileSystemOptions,
+  FileSystemUndeleteResponse
 } from "./models";
 import { Pipeline, StoragePipelineOptions, newPipeline } from "./Pipeline";
 import { StorageClient } from "./StorageClient";
@@ -356,5 +360,103 @@ export class DataLakeServiceClient extends StorageClient {
     ).toString();
 
     return appendToURLQuery(this.url, sas);
+  }
+
+  /**
+   * Renames an existing File System.
+   *
+   * @param {string} sourceFileSystemName The name of the source File System.
+   * @param {string} destinationContainerName The new name of the File System.
+   * @param {ServiceRenameFileSystemOptions} [options] Options to configure File System Rename operation.
+   * @memberof DataLakeServiceClient
+   */
+  // @ts-ignore Need to hide this interface for now. Make it public and turn on the live tests for it when the service is ready.
+  private async renameFileSystem(
+    sourceFileSystemName: string,
+    destinationFileSystemName: string,
+    options: ServiceRenameFileSystemOptions = {}
+  ): Promise<{
+    fileSystemClient: DataLakeFileSystemClient;
+    fileSystemRenameResponse: FileSystemRenameResponse;
+  }> {
+    const { span, spanOptions } = createSpan(
+      "DataLakeServiceClient-renameFileSystem",
+      options.tracingOptions
+    );
+    try {
+      // const res = await this.blobServiceClient.renameContainer(
+      const res = await this.blobServiceClient["renameContainer"](
+        sourceFileSystemName,
+        destinationFileSystemName,
+        {
+          ...options,
+          tracingOptions: { ...options.tracingOptions, spanOptions }
+        }
+      );
+
+      const fileSystemClient = this.getFileSystemClient(destinationFileSystemName);
+      return {
+        fileSystemClient,
+        fileSystemRenameResponse: res.containerRenameResponse
+      };
+    } catch (e) {
+      span.setStatus({
+        code: CanonicalCode.UNKNOWN,
+        message: e.message
+      });
+      throw e;
+    } finally {
+      span.end();
+    }
+  }
+
+  /**
+   * Restore a previously deleted File System.
+   * This API is only functional if Container Soft Delete is enabled for the storage account.
+   *
+   * @param {string} deletedFileSystemName The name of the source File System.
+   * @param {string} deleteFileSystemVersion The new name of the File System.
+   * @param {ServiceRenameFileSystemOptions} [options] Options to configure File System Restore operation.
+   * @memberof DataLakeServiceClient
+   */
+  public async undeleteFileSystem(
+    deletedFileSystemName: string,
+    deleteFileSystemVersion: string,
+    options: ServiceUndeleteFileSystemOptions = {}
+  ): Promise<{
+    fileSystemClient: DataLakeFileSystemClient;
+    fileSystemUndeleteResponse: FileSystemUndeleteResponse;
+  }> {
+    const { span, spanOptions } = createSpan(
+      "DataLakeServiceClient-undeleteFileSystem",
+      options.tracingOptions
+    );
+    try {
+      const res = await this.blobServiceClient.undeleteContainer(
+        deletedFileSystemName,
+        deleteFileSystemVersion,
+        {
+          ...options,
+          destinationContainerName: options.destinationFileSystemName,
+          tracingOptions: { ...options.tracingOptions, spanOptions }
+        }
+      );
+
+      const fileSystemClient = this.getFileSystemClient(
+        options.destinationFileSystemName || deletedFileSystemName
+      );
+      return {
+        fileSystemClient,
+        fileSystemUndeleteResponse: res.containerUndeleteResponse
+      };
+    } catch (e) {
+      span.setStatus({
+        code: CanonicalCode.UNKNOWN,
+        message: e.message
+      });
+      throw e;
+    } finally {
+      span.end();
+    }
   }
 }