+++ sdk/storage/storage-blob/CHANGELOG.md
@@ -1,7 +1,8 @@
 # Release History
 
-## 12.4.2 (Unreleased)
+## 12.5.0-beta.1 (2021-02-09)
 
+- Now support Batch operations scoped to the Container level. You can use `ContainerClient.getBlobBatchClient()` to get such a `BlobBatchClient`.
 
 ## 12.4.1 (2021-02-03)